Massive bot overhaul
We are no longer based on Red, we are now our own independent bot!

I'm leaving in the LICENSE file, 'cause that seems important, but other than that I don't think there was much else to keep. But tell me if I'm wrong!

(and yes, I updated the .gitignore to ignore the data stuff. The only thing from 'data' it includes is data/config.py, which is just the command prefix, the bot commander role name and whether the bot should DM you with command names.)