editing incorrect commands support
Now, you can edit incorrect commands (or commands that produce an error) within 20 seconds of sending it to make the bot do the command correctly!

At the moment of an incorrect command, the bot reacts with a red question mark. If your command is still invalid after 20 seconds, the bot will remove its reaction and do nothing else.

Oh also, I tried to make a console input thing? But it doesn't work, so ignore it